cockroachdb-jdbc-driver/src/test/java/io/cockroachdb/jdbc/util/AssertTest.java
package io.cockroachdb.jdbc.util;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Tag;
import org.junit.jupiter.api.Test;

@Tag("unit-test")
public class AssertTest {
    @Test
    public void whenAssertingConditions_expectConditionsToHold() {
        Assertions.assertDoesNotThrow(() -> Assert.notNull(1, "null!"));
        Assertions.assertDoesNotThrow(() -> Assert.isTrue(true, "!"));
        Assertions.assertDoesNotThrow(() -> Assert.hasText("a", "!"));
        Assertions.assertThrows(IllegalArgumentException.class, () -> Assert.notNull(null, "!"));
        Assertions.assertThrows(IllegalArgumentException.class, () -> Assert.isTrue(false, "!"));
        Assertions.assertThrows(IllegalArgumentException.class, () -> Assert.hasText("", "!"));
    }
}
